Обновление механизма spring security

+++ server/src/main/java/com/test/repos/UsersRepo.java
@@ -3,6 +3,6 @@ package com.test.repos;
 import com.test.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UsersRepo extends JpaRepository<User, Long> {
+public interface UsersRepo extends JpaRepository<User, Integer> {
     User findByUsername(String username);
 }
